Use better `assert_no_rvs` from logprob submodule
This utility can find RVs in inner graphs
